Recreate the data cleaning/org in R
This version of the a01_Fit_GLMMs recreates all of the work in R from raw data in CSV files for reproducibility purposes.

There are slightly marginal differences but the analysis remains unchanged, `mod4` still works best and `total_rain` is significant at p < 0.5, but only just. This may be due to handling of floating points?